Removed misleading sentence about FSMs.
In the section “State transitions and conflicts” you are writing that blockchains are essentially “state machines” and link to the Wikipedia article on finite-state machines (=finite-state automata). Now, while both Ethereum and Bitcoin are stateful and come with their own state transition functions, Ethereum in particular is not equivalent to any finite state automata. For example, a finite state machine cannot recognize the language {a^n b^n | n ≥ 1}, but Ethereum is Turing-complete and has no problem with this task (one can make an argument that Ethereum is not Turing-complete due to gas limits, but this can be compared to finite memory in traditional PC’s and is not relevant here). I think this point is important, since you are writing about smart contracts a bit later, which of course are heavily reliant on this property of the underlying blockchain. Therefore, I decided to remove the first sentence. Also, note that having a finite number of states is not the same as being a finite-state machine (the later being a strictly-defined technical term).