Added diagram depicting state transition.
I think it is good practice to illustrate technical matters with helpful diagrams. This not only reduces the time needed for the reader to become familiar with the subject, but also makes the text as a whole more readable. Disclainer: I borrowed this diagram from ethereum.org. It only serves as a proof of concept here. In a real world scenario, I would create this diagram myself and tailor it a bit more to the text.